fix: 퍼센트 계산 부동소수로 계산 후 정수로 Casting하도록 수정
* viewFriendsFundingDetail
* 퍼센트 계산에서 부동소수로 계산한 후 다시 정수로 casting되도록 수정

+++ src/main/java/kcs/funding/fundingboost/domain/service/FundingService.java
@@ -140,7 +140,7 @@ public class FundingService {
                     .map(ContributorDto::fromEntity)
                     .toList();
 
-            int contributedPercent = funding.getCollectPrice() / funding.getTotalPrice() * 100;
+            int contributedPercent = (int) ((float) funding.getCollectPrice() / funding.getTotalPrice() * 100);
 
             return FriendFundingDetailDto.fromEntity(friendFundingItemList, funding, contributorList,
                     contributedPercent);
